ml-model/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0

firebase-admin==6.0.1
python-dotenv==1.0.0

# Reddit API
praw==7.7.0

numpy
pandas
scipy
scikit-learn
aiohttp

# NLP + ML stack
nltk==3.8.1
regex==2023.6.3
tqdm==4.66.1

# HuggingFace transformers
transformers==4.32.1
torch==2.2.2

# Plotting
matplotlib==3.8.0

# For FastAPI forms
python-multipart==0.0.6
